delete blackgirlmagic read me
removed readme. conflicting with the pull request merge. 

"FCC-Tribute-Page---Black-Girl-Magic

A tribute page for Black Girl Magic, written to fulfill coursework for Free Code Camp. This is the first webpage that I've written and designed, so please feel free to provide feedback.

It can also be located on CodePen, where it was originally created. Here: https://codepen.io/YugHug/pen/BKQzJZ"